+++ .prettierrc
@@ -14,6 +14,8 @@
     "",
     "^(@shared)(/lib/.*)$",
     "",
+    "^(@shared)(/parsers/.*)$",
+    "",
     "^(@shared)(/types/.*)$",
     "",
     "^(@shared)(/icons/.*)$",
