+++ packages/shared/src/types/feedItems.types.ts
@@ -1,24 +1,14 @@
 import type {FieldValue} from 'firebase/firestore';
 import {z} from 'zod';
 
-import {
-  parseZodResult,
-  prefixErrorResult,
-  prefixResultIfError,
-} from '@shared/lib/errorUtils.shared';
-import {assertNever, makeId} from '@shared/lib/utils.shared';
+import {makeId} from '@shared/lib/utils.shared';
 
 import type {IconName} from '@shared/types/icons.types';
-import type {Result} from '@shared/types/result.types';
-import {makeSuccessResult} from '@shared/types/result.types';
 import type {KeyboardShortcutId} from '@shared/types/shortcuts.types';
 import type {TagId} from '@shared/types/tags.types';
-import {parseUserId, UserIdSchema, type UserId} from '@shared/types/user.types';
+import {UserIdSchema, type UserId} from '@shared/types/user.types';
 import type {UserFeedSubscriptionId} from '@shared/types/userFeedSubscriptions.types';
-import {
-  parseUserFeedSubscriptionId,
-  UserFeedSubscriptionIdSchema,
-} from '@shared/types/userFeedSubscriptions.types';
+import {UserFeedSubscriptionIdSchema} from '@shared/types/userFeedSubscriptions.types';
 import type {BaseStoreItem, Timestamp} from '@shared/types/utils.types';
 
 /**
@@ -26,7 +16,17 @@ import type {BaseStoreItem, Timestamp} from '@shared/types/utils.types';
  */
 export type FeedItemId = string & {readonly __brand: 'FeedItemIdBrand'};
 
+/**
+ * Zod schema for a {@link FeedItemId}.
+ */
 export const FeedItemIdSchema = z.string().uuid();
+
+/**
+ * Creates a new random {@link FeedItemId}.
+ */
+export function makeFeedItemId(): FeedItemId {
+  return makeId() as FeedItemId;
+}
 
 // TODO: Do I want to persist this or just compute it on the client?
 export enum FeedItemType {
@@ -53,32 +53,35 @@ export enum TriageStatus {
   Trashed = 'TRASHED',
 }
 
-const AppFeedItemSourceSchema = z.object({
+export const AppFeedItemSourceSchema = z.object({
   type: z.literal(FeedItemSourceType.App),
 });
 
-const ExtensionFeedItemSourceSchema = z.object({
+export const ExtensionFeedItemSourceSchema = z.object({
   type: z.literal(FeedItemSourceType.Extension),
 });
 
-const RssFeedItemSourceSchema = z.object({
+export const RssFeedItemSourceSchema = z.object({
   type: z.literal(FeedItemSourceType.RSS),
   userFeedSubscriptionId: UserFeedSubscriptionIdSchema,
 });
 
-export const FeedItemSourceSchema = z.union([
+export const FeedItemSourceSchema = z.discriminatedUnion('type', [
   AppFeedItemSourceSchema,
   ExtensionFeedItemSourceSchema,
   RssFeedItemSourceSchema,
 ]);
 
+/**
+ * Zod schema for a {@link FeedItem}. This should be kept in sync with that interface.
+ */
 export const FeedItemSchema = z.object({
   feedItemId: FeedItemIdSchema,
   userId: UserIdSchema,
   type: z.nativeEnum(FeedItemType),
   source: FeedItemSourceSchema,
   url: z.string(),
-  title: z.string(),
+  // title: z.string(),
   description: z.string(),
   outgoingLinks: z.array(z.string()),
   triageStatus: z.nativeEnum(TriageStatus),
@@ -87,122 +90,12 @@ export const FeedItemSchema = z.object({
   createdTime: z.string().datetime(),
   lastUpdatedTime: z.string().datetime(),
 });
-
-/**
- * Parses a {@link FeedItemId} from a plain string. Returns an `ErrorResult` if the string is not
- * valid.
- */
-export function parseFeedItemId(maybeFeedItemId: string): Result<FeedItemId> {
-  const parsedResult = parseZodResult(FeedItemIdSchema, maybeFeedItemId);
-  if (!parsedResult.success) {
-    return prefixErrorResult(parsedResult, 'Invalid feed item ID');
-  }
-  return makeSuccessResult(parsedResult.value as FeedItemId);
-}
-
-/**
- * Creates a new random {@link FeedItemId}.
- */
-export function makeFeedItemId(): FeedItemId {
-  return makeId() as FeedItemId;
-}
-
-/**
- * Parses a {@link FeedItemSource} from an unknown value. Returns an `ErrorResult` if the value is
- * not valid.
- */
-export function parseFeedItemSource(
-  source: unknown,
-  sourceType: FeedItemSourceType
-): Result<FeedItemSource> {
-  switch (sourceType) {
-    case FeedItemSourceType.App:
-      return parseAppFeedItemSource(source);
-    case FeedItemSourceType.Extension:
-      return parseExtensionFeedItemSource(source);
-    case FeedItemSourceType.RSS:
-      return parseRssFeedItemSource(source);
-    default:
-      assertNever(sourceType);
-  }
-}
-
-function parseAppFeedItemSource(source: unknown): Result<FeedItemAppSource> {
-  const parsedResult = parseZodResult(AppFeedItemSourceSchema, source);
-  if (!parsedResult.success) {
-    return prefixErrorResult(parsedResult, 'Invalid app feed item source');
-  }
-  return makeSuccessResult(parsedResult.value);
-}
-
-function parseExtensionFeedItemSource(source: unknown): Result<FeedItemExtensionSource> {
-  const parsedResult = parseZodResult(ExtensionFeedItemSourceSchema, source);
-  if (!parsedResult.success) {
-    return prefixErrorResult(parsedResult, 'Invalid extension feed item source');
-  }
-  return makeSuccessResult(parsedResult.value);
-}
-
-function parseRssFeedItemSource(source: unknown): Result<FeedItemRSSSource> {
-  const parsedResult = parseZodResult(RssFeedItemSourceSchema, source);
-  if (!parsedResult.success) {
-    return prefixErrorResult(parsedResult, 'Invalid RSS feed item source');
-  }
-  const parsedUserFeedSubscriptionIdResult = parseUserFeedSubscriptionId(
-    parsedResult.value.userFeedSubscriptionId
-  );
-  if (!parsedUserFeedSubscriptionIdResult.success) return parsedUserFeedSubscriptionIdResult;
-
-  return makeSuccessResult({
-    type: FeedItemSourceType.RSS,
-    userFeedSubscriptionId: parsedUserFeedSubscriptionIdResult.value,
-  });
-}
-
-/**
- * Parses a {@link FeedItem} from an unknown value. Returns an `ErrorResult` if the value is not
- * valid.
- */
-export function parseFeedItem(maybeFeedItem: unknown): Result<FeedItem> {
-  const parsedFeedItemResult = parseZodResult(FeedItemSchema, maybeFeedItem);
-  if (!parsedFeedItemResult.success) {
-    return prefixResultIfError(parsedFeedItemResult, 'Invalid feed item');
-  }
-
-  const parsedIdResult = parseFeedItemId(parsedFeedItemResult.value.feedItemId);
-  if (!parsedIdResult.success) return parsedIdResult;
-
-  const parsedUserIdResult = parseUserId(parsedFeedItemResult.value.userId);
-  if (!parsedUserIdResult.success) return parsedUserIdResult;
-
-  const parsedSourceResult = parseFeedItemSource(parsedFeedItemResult.value.source);
-  if (!parsedSourceResult.success) return parsedSourceResult;
-
-  const {url, title, createdTime, lastUpdatedTime} = parsedFeedItemResult.value;
-
-  return makeSuccessResult({
-    type: parsedFeedItemResult.value.type,
-    userId: parsedUserIdResult.value,
-    source: parsedSourceResult.value,
-    feedItemId: parsedIdResult.value,
-    url,
-    title,
-    description: '',
-    outgoingLinks: [],
-    triageStatus: parsedFeedItemResult.value.triageStatus,
-    tagIds: {},
-    lastImportedTime: new Date(parsedFeedItemResult.value.lastImportedTime) as unknown as Timestamp,
-    createdTime: new Date(createdTime) as unknown as Timestamp,
-    lastUpdatedTime: new Date(lastUpdatedTime) as unknown as Timestamp,
-  });
-}
-
 interface BaseFeedItemSource {
   // TODO: Consider renaming this to `sourceType`.
   readonly type: FeedItemSourceType;
 }
 
-interface FeedItemAppSource extends BaseFeedItemSource {
+export interface FeedItemAppSource extends BaseFeedItemSource {
   readonly type: FeedItemSourceType.App;
 }
 
@@ -210,7 +103,7 @@ export const FEED_ITEM_EXTENSION_SOURCE: FeedItemExtensionSource = {
   type: FeedItemSourceType.Extension,
 };
 
-interface FeedItemExtensionSource extends BaseFeedItemSource {
+export interface FeedItemExtensionSource extends BaseFeedItemSource {
   readonly type: FeedItemSourceType.Extension;
 }
 
@@ -218,7 +111,7 @@ export const FEED_ITEM_APP_SOURCE: FeedItemAppSource = {
   type: FeedItemSourceType.App,
 };
 
-interface FeedItemRSSSource extends BaseFeedItemSource {
+export interface FeedItemRSSSource extends BaseFeedItemSource {
   readonly type: FeedItemSourceType.RSS;
   readonly userFeedSubscriptionId: UserFeedSubscriptionId;
 }
@@ -232,10 +125,7 @@ export function makeFeedItemRSSSource(
   };
 }
 
-export type FeedItemSource =
-  | typeof FEED_ITEM_APP_SOURCE
-  | typeof FEED_ITEM_EXTENSION_SOURCE
-  | FeedItemRSSSource;
+export type FeedItemSource = FeedItemAppSource | FeedItemExtensionSource | FeedItemRSSSource;
 
 interface BaseFeedItem extends BaseStoreItem {
   readonly feedItemId: FeedItemId;
